\name{DataFrameFromJson}
\alias{DataFrameFromJson}
\title{DataFrameFromJson}
\usage{
DataFrameFromJson(json.objects, column.classes)
}
\arguments{
  \item{json.objects}{a list of json objects from fromJSON}

  \item{column.classes}{a named list of columns and
  converter}
}
\description{
Converts a list of JSON objects to a data.frame
}

