\name{FormatFreeAgentImport}
\alias{FormatFreeAgentImport}
\title{Formats a list of transactions as a CSV
file ready to import into FreeAgent}
\usage{
FormatFreeAgentImport(file, date, amount, description)
}
\description{
Formats a list of transactions as a CSV file ready to
import into FreeAgent
}

