\name{GetFreeAgentBankTransactions}
\alias{GetFreeAgentBankTransactions}
\title{Retrieves the list of transactions stored in FreeAgent for a given bank account}
\usage{
GetFreeAgentBankTransactions(bank.account, ...)
}
\description{
Retrieves the list of transactions stored in FreeAgent for
a given bank account
}

