\name{GetFreeAgentInvoices}
\alias{GetFreeAgentInvoices}
\title{Retrieves invoices from FreeAgent}
\usage{
GetFreeAgentInvoices(...)
}
\arguments{
  \item{view}{Can be recent_open_or_overdue,
  open_or_overdue, draft, scheduled_to_email,
  thank_you_emails, reminder_emails, last_N_months}
}
\description{
Retrieves invoices from FreeAgent
}

