\name{GetFreeAgentResources}
\alias{GetFreeAgentResources}
\title{Fetches a list of FreeAgent resources, optionally requesting additional
pages until the list is complete}
\usage{
GetFreeAgentResources(resource.name, wrapper.name = resource.name,
  fetch.all = TRUE, ...)
}
\description{
Fetches a list of FreeAgent resources, optionally
requesting additional pages until the list is complete
}

