\name{PostFreeAgent}
\alias{PostFreeAgent}
\title{Fetches a list of FreeAgent resources, optionally requesting additional
pages until the list is complete}
\usage{
PostFreeAgent(resource.name, ...)
}
\description{
Fetches a list of FreeAgent resources, optionally
requesting additional pages until the list is complete
}

