\name{UploadFreeAgentStatement}
\alias{UploadFreeAgentStatement}
\title{Formats a list of transactions as a CSV
file ready to import into FreeAgent}
\usage{
UploadFreeAgentStatement(account, date, amount, description)
}
\description{
Formats a list of transactions as a CSV file ready to
import into FreeAgent
}

